dom/tests/mochitest/pointerlock/test_suppressSomeMouseEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_cursorPosEvents.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #child {
      width: 100px;
      height: 100px;
      background-color:Green;
    }
  </style>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>

  <div id="parent">
    <div id="child"></div>
  </div>

  <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Test will check to make sure that the following mouse events are no
       * longer executed in pointer lock.
       *  - mouseover, mouseout, mouseenter, mouseleave
       */

      SpecialPowers.setBoolPref("full-screen-api.enabled", true);
      SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                 false);

      SimpleTest.waitForExplicitFinish();

      function PointerEventStats() {
        this.mouseEnter = false;
        this.mouseLeave = false;
        this.mouseOver = false;
        this.mouseOut = false;
        this.drag = false;
        this.dragStart = false;
        this.dragEnter = false;
        this.dragOver = false;
        this.dragLeave = false;
        this.dragEnd = false;
        this.drop = false;
      }

      var parent
        , child
        , parentStats = new PointerEventStats()
        , childStats = new PointerEventStats()
        , isPointerLocked = false;

      function runTests () {
        ok(isPointerLocked, "expected mouse to be locked, but wasn't.");

        is(childStats.mouseEnter, false,
           "child's mouseenter should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseOver, false,
           "child's mouseover should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseLeave, false,
           "child's mouseleave should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseOut, false,
           "child's mouseout should not be firing in Full Screen and Pointer Lock.");
        is(childStats.drag, false,
           "child's drag should not be firing in Full Screen and Pointer Lock.");
        is(childStats.dragStart, false,
           "child's dragStart should not be firing in Full Screen and Pointer Lock.");
        is(childStats.dragEnter, false,
           "child's dragEnter should not be firing in Full Screen and Pointer Lock.");
        is(childStats.dragOver, false,
           "child's dragOver should not be firing in Full Screen and Pointer Lock.");
        is(childStats.dragLeave, false,
           "child's dragLeave should not be firing in Full Screen and Pointer Lock.");
        is(childStats.dragEnd, false,
           "child's dragEnd should not be firing in Full Screen and Pointer Lock.");
        is(childStats.drop, false,
           "child's drop should not be firing in Full Screen and Pointer Lock.");


        is(parentStats.mouseEnter, false,
           "parent's mouseenter should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseOver, false,
           "parent's mouseover should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseLeave, false,
           "parent's mouseleave should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseOut, false,
           "parent's mouseout should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.drag, false,
           "parent's drag should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.dragStart, false,
           "parent's dragStart should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.dragEnter, false,
           "parent's dragEnter should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.dragOver, false,
           "parent's dragOver should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.dragLeave, false,
          "parent's dragLeave should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.dragEnd, false,
           "parent's dragEnd should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.drop, false,
           "parent's drop should not be firing in Full Screen and Pointer Lock.");
      }

      var parentMoveListener = function () {
        isPointerLocked = !!document.mozPointerLockElement;
        removeEventListeners();
        document.mozExitPointerLock();
      };

      var parentOutListener = function (e) {
          parentStats.mouseOut = true;
      };
      var parentLeaveListener = function (e) {
          parentStats.mouseLeave = true;
      };
      var parentOverListener = function (e) {
          parentStats.mouseOver = true;
      };
      var parentEnterListener = function (e) {
          parentStats.mouseEnter = true;
      };
      var parentDragListener = function (e) {
          parentStats.drag = true;
      };
      var parentDragStartListener = function (e) {
          parentStats.dragStart = true;
      };
      var parentDragEnterListener = function (e) {
          parentStats.dragEnter = true;
      };
      var parentDragOverListener = function (e) {
          parentStats.dragOver = true;
      };
      var parentDragLeaveListener = function (e) {
          parentStats.dragLeave = true;
      };
      var parentDragEndListener = function (e) {
          parentStats.dragEnd = true;
      };
      var parentDropListener = function (e) {
          parentStats.drop = true;
      };

      var childOutListener = function (e) {
          childStats.mouseOut = true;
      };
      var childLeaveListener = function (e) {
          childStats.mouseLeave = true;
      };
      var childOverListener = function (e) {
          childStats.mouseOver = true;
      };
      var childEnterListener = function (e) {
          childStats.mouseEnter = true;
      };
      var childDragListener = function (e) {
          childStats.drag = true;
      };
      var childDragStartListener = function (e) {
          childStats.dragStart = true;
      };
      var childDragEnterListener = function (e) {
          childStats.dragEnter = true;
      };
      var childDragOverListener = function (e) {
          childStats.dragOver = true;
      };
      var childDragLeaveListener = function (e) {
          childStats.dragLeave = true;
      };
      var childDragEndListener = function (e) {
          childStats.dragEnd = true;
      };
      var childDropListener = function (e) {
          childStats.drop = true;
      };

      function addEventListeners() {
        parent.addEventListener("mousemove", parentMoveListener, false);

        parent.addEventListener("mouseout", parentOutListener, false);
        parent.addEventListener("mouseleave", parentLeaveListener, false);
        parent.addEventListener("mouseover", parentOverListener, false);
        parent.addEventListener("mouseenter", parentEnterListener, false);
        parent.addEventListener("drag", parentDragListener, false);
        parent.addEventListener("dragstart", parentDragStartListener, false);
        parent.addEventListener("dragenter", parentDragEnterListener, false);
        parent.addEventListener("dragover", parentDragOverListener, false);
        parent.addEventListener("dragleave", parentDragLeaveListener, false);
        parent.addEventListener("dragend", parentDragEndListener, false);
        parent.addEventListener("drop", parentDropListener, false);

        child.addEventListener("mouseout", childOutListener, false);
        child.addEventListener("mouseleave", childLeaveListener, false);
        child.addEventListener("mouseover", childOverListener, false);
        child.addEventListener("mouseenter", childEnterListener, false);
        child.addEventListener("drag", childDragListener, false);
        child.addEventListener("dragstart", childDragStartListener, false);
        child.addEventListener("dragenter", childDragEnterListener, false);
        child.addEventListener("dragover", childDragOverListener, false);
        child.addEventListener("dragleave", childDragLeaveListener, false);
        child.addEventListener("dragend", childDragEndListener, false);
        child.addEventListener("drop", childDropListener, false);
      }

      function removeEventListeners() {
        parent.removeEventListener("mousemove", parentMoveListener, false);

        parent.removeEventListener("mouseout", parentOutListener, false);
        parent.removeEventListener("mouseleave", parentLeaveListener, false);
        parent.removeEventListener("mouseover", parentOverListener, false);
        parent.removeEventListener("mouseenter", parentEnterListener, false);
        parent.removeEventListener("drag", parentDragListener, false);
        parent.removeEventListener("dragstart", parentDragStartListener, false);
        parent.removeEventListener("dragenter", parentDragEnterListener, false);
        parent.removeEventListener("dragover", parentDragOverListener, false);
        parent.removeEventListener("dragleave", parentDragLeaveListener, false);
        parent.removeEventListener("dragend", parentDragEndListener, false);
        parent.removeEventListener("drop", parentDropListener, false);


        child.removeEventListener("mouseout", childOutListener, false);
        child.removeEventListener("mouseleave", childLeaveListener, false);
        child.removeEventListener("mouseover"  , childOverListener, false);
        child.removeEventListener("mouseenter", childEnterListener, false);
        child.removeEventListener("drag", childDragListener, false);
        child.removeEventListener("dragstart", childDragStartListener, false);
        child.removeEventListener("dragenter", childDragEnterListener, false);
        child.removeEventListener("dragover", childDragOverListener, false);
        child.removeEventListener("dragleave", childDragLeaveListener, false);
        child.removeEventListener("dragend", childDragEndListener, false);
        child.removeEventListener("drop", childDropListener, false);
      }

      function dispatchDragEvent(eventName) {
        var event = document.createEvent("DragEvents");
        event.initDragEvent(eventName, true, true, window, 0, 5, 5, 5, 5,
                            false, false, false, false, 0, null, null);
        child.dispatchEvent(event);
      }

      document.addEventListener("mozpointerlockchange", function (e) {
        if (document.mozPointerLockElement === parent) {
          addEventListeners();
          dispatchDragEvent("dragstart");
          dispatchDragEvent("drag");
          dispatchDragEvent("dragenter");
          dispatchDragEvent("dragover");
          dispatchDragEvent("dragleave");
          dispatchDragEvent("drop");
          dispatchDragEvent("dragend");
          synthesizeMouseAtCenter(child, { type: "mousemove" }, window);
        }
        else {
          document.mozCancelFullScreen();
        }
      }, false);

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreenElement === parent) {
          parent.mozRequestPointerLock();
        }
        else {
          runTests();
          SimpleTest.finish();
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          parent = document.getElementById("parent");
          child = document.getElementById("child");
          parent.mozRequestFullScreen();
        });
      }
  </script>
</body>
</html>
